app/requirements.txt
aiohttp==3.10.9
alembic==1.13.3
asyncpg==0.29.0
et-xmlfile==1.1.0
fastapi==0.115.0
PyJWT==2.9.0
numpy==2.1.2
pydantic==2.9.2
pydantic_core==2.23.4
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pwdlib[argon2]==0.3.0
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.14.1
SQLAlchemy==2.0.35
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
starlette==0.38.6
statsmodels==0.14.4
typing_extensions==4.12.2
uvicorn==0.32.0